pyproject.toml
[project]
name = "templates-static"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.100.0",
    "generativeai>=0.0.1",
    "jinja2>=3.1.6",
    "pydantic>=2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "uvicorn==0.22.0",
]
